В классе RepositoryImpl.kt добавлена функция для получения данных с сервера о выбранном фильме.

+++ app/src/main/java/ru/profitsw2000/moviecollectiondb/model/representation_tmdb/DescriptionDTO.kt
@@ -11,5 +11,5 @@ data class DescriptionDTO(
     val revenue: Int?,
     val runtime: Int?,
     val title: String?,
-    val vote_average: Int?
+    val vote_average: Float?
 )
